requirements.txt
# Core APM
elastic-apm>=6.15.1

# Web Frameworks
flask>=2.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP Client
requests>=2.31.0
httpx>=0.25.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
